Form storage mv from local storage to App.js state. Rename stages.
