+++ src/stages/stage3-verify.js
