+++ src/stages/stage4-register.js
