+++ src/stages/stage6-final.js
